Move dynamic shapes into core
