+++ org.mwc.debrief.core/META-INF/MANIFEST.MF
@@ -35,7 +35,6 @@ Export-Package: com.topografix.gpx.v10,
  org.mwc.debrief.core.editors.painters.snail,
  org.mwc.debrief.core.gpx,
  org.mwc.debrief.core.gpx.mappers,
- org.mwc.debrief.core.handlers,
  org.mwc.debrief.core.interfaces,
  org.mwc.debrief.core.loaders,
  org.mwc.debrief.core.loaders.xml_handlers,
